Rename holiday package to publicholiday
